Get user from Session instead of hardcoded value

+++ src/api/private/tokens/tokens.controller.ts
@@ -11,14 +11,19 @@ import {
   HttpCode,
   Param,
   Post,
+  UnauthorizedException,
+  UseGuards,
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 
 import { AuthTokenWithSecretDto } from '../../../auth/auth-token-with-secret.dto';
 import { AuthTokenDto } from '../../../auth/auth-token.dto';
 import { AuthService } from '../../../auth/auth.service';
+import { SessionGuard } from '../../../identity/session.guard';
 import { ConsoleLoggerService } from '../../../logger/console-logger.service';
+import { User } from '../../../users/user.entity';
 import { TimestampMillis } from '../../../utils/timestamp';
+import { RequestUser } from '../../utils/request-user.decorator';
 
 @ApiTags('tokens')
 @Controller('tokens')
@@ -30,30 +35,43 @@ export class TokensController {
     this.logger.setContext(TokensController.name);
   }
 
+  @UseGuards(SessionGuard)
   @Get()
-  async getUserTokens(): Promise<AuthTokenDto[]> {
-    // ToDo: Get real userName
-    return (await this.authService.getTokensByUsername('hardcoded')).map(
+  async getUserTokens(@RequestUser() user: User): Promise<AuthTokenDto[]> {
+    return (await this.authService.getTokensByUsername(user.userName)).map(
       (token) => this.authService.toAuthTokenDto(token),
     );
   }
 
+  @UseGuards(SessionGuard)
   @Post()
   async postTokenRequest(
     @Body('label') label: string,
     @Body('validUntil') validUntil: TimestampMillis,
+    @RequestUser() user: User,
   ): Promise<AuthTokenWithSecretDto> {
-    // ToDo: Get real userName
     return await this.authService.createTokenForUser(
-      'hardcoded',
+      user.userName,
       label,
       validUntil,
     );
   }
 
+  @UseGuards(SessionGuard)
   @Delete('/:keyId')
   @HttpCode(204)
-  async deleteToken(@Param('keyId') keyId: string): Promise<void> {
-    return await this.authService.removeToken(keyId);
+  async deleteToken(
+    @RequestUser() user: User,
+    @Param('keyId') keyId: string,
+  ): Promise<void> {
+    const tokens = await this.authService.getTokensByUsername(user.userName);
+    for (const token of tokens) {
+      if (token.keyId == keyId) {
+        return await this.authService.removeToken(keyId);
+      }
+    }
+    throw new UnauthorizedException(
+      'User is not authorized to delete this token',
+    );
   }
 }
